requirements.txt
pytest
pytest-playwright
pytest-xdist
allure-pytest
pytest-rerunfailures
